+++ include/hc_printf.hpp
@@ -311,100 +311,4 @@ static inline PrintfError printf(const char* format_string, All... all) [[hc,cpu
 
 #endif
 
-// regex for finding format string specifiers
-static const std::regex specifierPattern("(%){1}[-+#0]*[0-9]*((.)[0-9]+){0,1}([hl]*)([diuoxXfFeEgGaAcsp]){1}");
-static const std::regex signedIntegerPattern("(%){1}[-+#0]*[0-9]*((.)[0-9]+){0,1}([hl]*)([cdi]){1}");
-static const std::regex unsignedIntegerPattern("(%){1}[-+#0]*[0-9]*((.)[0-9]+){0,1}([hl]*)([uoxX]){1}");
-static const std::regex floatPattern("(%){1}[-+#0]*[0-9]*((.)[0-9]+){0,1}([fFeEgGaA]){1}");
-static const std::regex pointerPattern("(%){1}[ps]");
-static const std::regex doubleAmpersandPattern("(%){2}");
-static const std::string ampersand("%");
-
-static inline void processPrintfPackets(PrintfPacket* packets, const unsigned int numPackets) {
-
-  for (unsigned int i = 0; i < numPackets; ) {
-
-    unsigned int numPrintfArgs = packets[i++].data.ui;
-    if (numPrintfArgs == 0)
-      continue;
-
-    // get the format
-    unsigned int formatStringIndex = i++;
-    assert(packets[formatStringIndex].type == PRINTF_CHAR_PTR
-           || packets[formatStringIndex].type == PRINTF_CONST_CHAR_PTR);
-    std::string formatString((const char*)packets[formatStringIndex].data.cptr);
-    std::smatch specifierMatches;
-
-#if HC_PRINTF_DEBUG
-    std::printf("%s:%d \t number of matches = %d\n", __FUNCTION__, __LINE__, (int)specifierMatches.size());
-#endif
-
-    for (unsigned int j = 1; j < numPrintfArgs; ++j, ++i) {
-
-      if (!std::regex_search(formatString, specifierMatches, specifierPattern)) {
-        // More printf argument than format specifier??
-        // Just skip to the next printf request
-        i+=(numPrintfArgs - j);
-        break;
-      }
-
-      std::string specifier = specifierMatches.str();
-#if HC_PRINTF_DEBUG
-      std::cout << " (specifier found: " << specifier << ") ";
-#endif
-
-      // print the substring before the specifier
-      // clean up all the double ampersands
-      std::string prefix = specifierMatches.prefix();
-      prefix = std::regex_replace(prefix,doubleAmpersandPattern,ampersand);
-      std::printf("%s",prefix.c_str());
-
-      std::smatch specifierTypeMatch;
-      if (std::regex_search(specifier, specifierTypeMatch, unsignedIntegerPattern)) {
-        std::printf(specifier.c_str(), packets[i].data.ui);
-      } else if (std::regex_search(specifier, specifierTypeMatch, signedIntegerPattern)) {
-        std::printf(specifier.c_str(), packets[i].data.i);
-      } else if (std::regex_search(specifier, specifierTypeMatch, floatPattern)) {
-        if (packets[i].type == PRINTF_HALF)
-          std::cout << static_cast<float>(packets[i].data.h);
-        else if (packets[i].type == PRINTF_FLOAT)
-          std::printf(specifier.c_str(), packets[i].data.f);
-        else
-          std::printf(specifier.c_str(), packets[i].data.d);
-      } else if (std::regex_search(specifier, specifierTypeMatch, pointerPattern)) {
-        std::printf(specifier.c_str(), packets[i].data.cptr);
-      }
-      else {
-        assert(false);
-      }
-      formatString = specifierMatches.suffix();
-    }
-    // print the substring after the last specifier
-    // clean up all the double ampersands before printing
-    formatString = std::regex_replace(formatString,doubleAmpersandPattern,ampersand);
-    std::printf("%s",formatString.c_str());
-  }
-  std::flush(std::cout);
-}
-
-static inline void processPrintfBuffer(PrintfPacket* gpuBuffer) {
-
-  if (gpuBuffer == nullptr) return;
-
-  unsigned int cursor = gpuBuffer[PRINTF_OFFSETS].data.uia[0];
-
-  // check whether the printf buffer is non-empty
-  if (cursor !=  PRINTF_HEADER_SIZE) {
-    unsigned int bufferSize = gpuBuffer[PRINTF_BUFFER_SIZE].data.ui;
-    unsigned int numPackets = ((bufferSize<cursor)?bufferSize:cursor) - PRINTF_HEADER_SIZE;
-
-    processPrintfPackets(gpuBuffer+PRINTF_HEADER_SIZE, numPackets);
-
-    // reset the printf buffer and string buffer
-    gpuBuffer[PRINTF_OFFSETS].data.uia[0] = PRINTF_HEADER_SIZE;
-    gpuBuffer[PRINTF_OFFSETS].data.uia[1] = 0;
-  }
-}
-
-
 } // namespace hc
